requirements.txt
requests==2.21.0
PyGithub==1.43.7
tabulate==0.8.3